chore: bring exec_and_watch together
