+++ src/command/subgraph/introspect.rs
@@ -3,11 +3,11 @@ use saucer::{clap, Parser};
 use serde::Serialize;
 use std::collections::HashMap;
 
-use crate::{command::output::JsonOutput, options::IntrospectOpts};
+use crate::options::IntrospectOpts;
 
 use rover_client::{
     blocking::GraphQLClient,
-    operations::subgraph::introspect::{self, SubgraphIntrospectInput, SubgraphIntrospectResponse},
+    operations::subgraph::introspect::{self, SubgraphIntrospectInput},
 };
 
 use crate::command::RoverOutput;
@@ -25,12 +25,12 @@ impl Introspect {
             self.exec_and_watch(&client, json)?;
             Ok(RoverOutput::EmptySuccess)
         } else {
-            let response = self.exec(&client, true)?;
-            Ok(RoverOutput::Introspection(response.result))
+            let sdl = self.exec(&client, true)?;
+            Ok(RoverOutput::Introspection(sdl))
         }
     }
 
-    pub fn exec(&self, client: &Client, should_retry: bool) -> Result<SubgraphIntrospectResponse> {
+    pub fn exec(&self, client: &Client, should_retry: bool) -> Result<String> {
         let client = GraphQLClient::new(&self.opts.endpoint.to_string(), client.clone())?;
 
         // add the flag headers to a hashmap to pass along to rover-client
@@ -41,55 +41,12 @@ impl Introspect {
             }
         };
 
-        Ok(introspect::run(
-            SubgraphIntrospectInput { headers },
-            &client,
-            should_retry,
-        )?)
+        Ok(introspect::run(SubgraphIntrospectInput { headers }, &client, should_retry)?.result)
     }
 
     pub fn exec_and_watch(&self, client: &Client, json: bool) -> Result<RoverOutput> {
-        let mut last_result = None;
-        loop {
-            match self.exec(client, false) {
-                Ok(response) => {
-                    let sdl = response.result.to_string();
-                    let mut was_updated = true;
-                    if let Some(last) = last_result {
-                        if last == response.result {
-                            was_updated = false
-                        }
-                    }
-
-                    if was_updated {
-                        let output = RoverOutput::Introspection(sdl.to_string());
-                        if json {
-                            let _ = JsonOutput::from(output).print();
-                        } else {
-                            let _ = output.print();
-                        }
-                    }
-                    last_result = Some(sdl);
-                }
-                Err(error) => {
-                    let mut was_updated = true;
-                    let e = error.to_string();
-                    if let Some(last) = last_result {
-                        if last == e {
-                            was_updated = false;
-                        }
-                    }
-                    if was_updated {
-                        if json {
-                            let _ = JsonOutput::from(error).print();
-                        } else {
-                            let _ = error.print();
-                        }
-                    }
-                    last_result = Some(e);
-                }
-            }
-            std::thread::sleep(std::time::Duration::from_secs(1))
-        }
+        self.opts
+            .exec_and_watch(|| self.exec(client, false), json)?;
+        Ok(RoverOutput::EmptySuccess)
     }
 }
