+++ src/options/introspect.rs
@@ -2,7 +2,11 @@ use reqwest::Url;
 use saucer::{clap, Parser};
 use serde::{Deserialize, Serialize};
 
-use crate::utils::parsers::parse_header;
+use crate::{
+    command::{output::JsonOutput, RoverOutput},
+    utils::parsers::parse_header,
+    Result,
+};
 
 #[derive(Debug, Serialize, Deserialize, Parser)]
 pub struct IntrospectOpts {
@@ -23,4 +27,53 @@ pub struct IntrospectOpts {
     /// poll the endpoint, printing the introspection result if/when its contents change
     #[clap(long)]
     pub watch: bool,
+}
+
+impl IntrospectOpts {
+    pub fn exec_and_watch<F>(&self, exec_fn: F, json: bool) -> Result<RoverOutput>
+    where
+        F: Fn() -> Result<String>,
+    {
+        let mut last_result = None;
+        loop {
+            match exec_fn() {
+                Ok(sdl) => {
+                    let mut was_updated = true;
+                    if let Some(last) = last_result {
+                        if last == sdl {
+                            was_updated = false
+                        }
+                    }
+
+                    if was_updated {
+                        let output = RoverOutput::Introspection(sdl.to_string());
+                        if json {
+                            let _ = JsonOutput::from(output).print();
+                        } else {
+                            let _ = output.print();
+                        }
+                    }
+                    last_result = Some(sdl);
+                }
+                Err(error) => {
+                    let mut was_updated = true;
+                    let e = error.to_string();
+                    if let Some(last) = last_result {
+                        if last == e {
+                            was_updated = false;
+                        }
+                    }
+                    if was_updated {
+                        if json {
+                            let _ = JsonOutput::from(error).print();
+                        } else {
+                            let _ = error.print();
+                        }
+                    }
+                    last_result = Some(e);
+                }
+            }
+            std::thread::sleep(std::time::Duration::from_secs(1))
+        }
+    }
 }
